Updated for Maps 9.7.1
Google has again refactored how they request for the location settings
status through Play Services. Flipping the right bit seems to do the
job. It will not give us a pretty little “optimal location” toast
notification but it works just as well. Unfortunately the necessary
class definition is in the second dex file and must be accessed after
that dex is loaded.